Refactor : line 차트들 코드 구조 변경, 사용하지 않는 로직 및 코드 제거
