+++ src/jetty-chart/src/components/x-axis/draw-x-axis-grid-line.js
@@ -59,7 +59,7 @@ export const DrawXAxisGridLine = ({
 
       return linePath;
     },
-    [horizontal, height, lineDashWidth, lineDashGap]
+    [horizontal, height, lineDashWidth, lineDashGap, lineDash, useAnimation]
   );
 
   if (!lineVisible) {
